Add files via upload
This is an interpreter for a Deterministic Finite Automata. The user can enter an input, and DFA file. The DFA file describes the rules of the DFA, and the input describes a language. The structure of dfa.txt will be as follows:
 Line 1: the states of the DFA (separated by commas, if there is more than one state)
 Line 2: the alphabet of the DFA (separated by commas, if there is more than one symbol)
 Line 3: the starting state of the DFA
 Line 4: the final/accept states of the DFA (separated by commas, if there is more than one accept state)
 Line 5 and onward: the transition rules, where each rule takes the form a,b,c (where being in state a and reading symbol b transitions to new state c)
After the program constructs a DFA, read in another file called input.txt, where each line in the file is a string that will run on your DFA.
Both dfa.txt and input.txt should be assumed to be in the same directory as your Python program.
This simulates each string in input.txt using your DFA, and writes the output (accept or reject) into a new file called output.txt (also in the same directory as your Python program, and each result on a new line).